+++ backend/src/services/auth.service.js
@@ -2,6 +2,7 @@ const bcrypt = require("bcrypt");
 const jwt = require("jsonwebtoken");
 const { createTenant } = require("../services/tenant.service")
 const { createUser } = require("../services/user.service")
+const User = require("../models/user.model");
 
 const signup = async ({ agencyName, email, password }) => {
 
@@ -31,4 +32,24 @@ const signup = async ({ agencyName, email, password }) => {
 return { token };
 };
 
-module.exports = { signup };
+const login = async ({ email, password }) => {
+    const user = await User.findOne({ email });
+    if(!user) throw new Error("Invalid credentials");
+
+    const isMatch = await bcrypt.compare(password, user.password);
+    if (!isMatch) throw new Error("Invalid credentials");
+
+    const token = jwt.sign(
+        {
+            userId: user._id,
+            tenantId: user.tenantId,
+            role: user.role
+        },
+        process.env.JWT_SECRET,
+        { expiresIn: "1d" }
+    );
+
+    return { token };
+}
+
+module.exports = { signup, login };